Independence of sequence and tuple operations.
Up until now, preprocessor sequences and tuples within the ppmpf framework
have been relying on counting macro arguments with a preset macro like
PPMPF_NARGS / PPMPF_VARGS for operations like empty checking. With the
changes just introduced, this is no longer the case. Following this
commit, the implementation of argument counting macros is to be revised
to take advantage of the PPMPF_EMPTY_ macro used together with a fold
operation.