Splitting base.hh macro constructs into more headers within ppmpf.
With the introduction of alu.hh, core.hh, tupseq.hh, a lot of macros that were
merged into the initial base.hh file have been removed from that and placed
into separate headers. The remaining base.hh macros consist in the original
PPMPF_REPEAT construct (up to 10000 elements in horizontal repeat) and macro
argument counting (which is to be revised in view of tupseq.hh functionality).