C99 ISO compliant count of __VA_ARGS__ arguments.
Using ppmpf tuple folding, we can calculate the number of arguments contained
in the __VA_ARGS__ builtin in a ISO C99 compliant way. Using PPMPF_IARGS(...)
returns a 4 member digit token ppmpf sequence encoding the number of arguments
passed in the ppmpf representation of integers format, while PPMPF_NARGS(...)
does the same thing but returns a regular, readily readable integer. The
maximum number of arguments that can be counted is 9999. If no arguments are
passed to either of these counters, the result is zero. These two function
macros deprecate the earlier versions in odreex/ppmpf/base.hh that had a
limit of 128 arguments. The method works in C++11 standard compliant
compilers (e.g. g++, clang++) and passes the -pedantic checks.