C99 ISO macro arguments counter fix, other fixes.
When the last element in a ppmpf tuple is a token corresponding to a macro,
the checks involved in the counting algorithm erroneously allowed that macro
to expand. Thus, rewiring was necessary in the empty preprocessor token
detection cascades. Other fixes result in a more homogeneous naming convention
for PPMPF_TUP_* and PPMPF_SEQ_* macro family operations.