ppmpf tuple / sequence items always enclosed during folding.
The change affects a series of macros already implemented and makes certain
that each item passed to the binary function macro that is used during
folding is enclosed by balanced parentheses, therefore covering the case
of processing a sequence made out of tuples as items being folded. Thus,
the "dereferencing" macro has no longer meaning as an argument in the
recursive macros making up the basis for the fold construct.