Several fixes related to folds, ppmpf collections and their functions.
Fixed a series of bugs related ppmpf collection pop / get functions, folds,
empty checks that would end up in specific use case scenarios and / or were
of temporary implementation in order for ppmpf to grow. Minimum and maximum
numbers remain 0 and 9999 respectively with PPMPF_IMINV_() and PPMPF_IMAXV_()
having switched places with PPMPF_IMINV() and PPMPF_IMAXV(). Other changes to
function macros like PPMPF_TUP_EMPTY(...), PPMPF_TUP_SIZEOF(...) and of course
PPMPF_SEQ_SIZEOF(...).

When a ppmpf collection is undergoing a fold, the binary function used
to perform that fold is given two macro arguments that are always enclosed
in balanced parentheses, so it is up to the end user to apply PPMPF_DREF(...)
where applicable, for the time being. This will change once a few other
features are added to the macro constructs responsible for folding.