Converting ppmpf tuples to sequences, initial ppmpf refactoring.
Given a ppmpf tuple of the form (a,b,c,d,e,f,...), using the PPMPF_TUP2SEQ we
perform the conversion to the equivalent (a)(b)(c)(d)(e)(f)... ppmpf sequence.

Functions related to specific operations that can be logically bundled into
distinct units, will be having their own directory within the ppmpf root, this
is a guideline for refactoring the layout of the ppmpf codebase.