Added PPMPF_TUP_MAP as a high order map macro for ppmpf tuples.
Given a ppmpf tuple of the form (1,2,3,...,n) and a function macro f, when
PPMPF_TUP_MAP is used, a tuple of the form (f(1), f(2), f(3), ... , f(n))
is the result of the the macro expansion sequence. Examples follow, where
-> is a symbol used for indicating full macro expansion. When f is defined
as well, it as well is expanded in the result.

    PPMPF_TUP_MAP(f,())          -> ()
    PPMPF_TUP_MAP(f,(token))     -> (f(token))
    PPMPF_TUP_MAP(f,(1,2,3,4,5)) -> (f(1),f(2),f(3),f(4),f(5))