Initial ppmpf tuple restructuring.
The current status quo of ppmpf was to have everything into a single, common
root folder. A more modular design in the headers involved is to be followed
with "essence.hh" headers being home for important, basic macros a high level
macro construct must have in order to work as advertised. In a way, essence.hh
headers within ppmpf folders are the home for core macro constructs and
cannot be included separately.