Extensive header reordering throughout ppmpf, charseq test added.
Most of the changes reflect ppmpf header relocation to a directory layout
compatible with the production - ready form for ppmpf, therefore the
ppmpf header modifications are header path changes. Some tests for the
basic constructs in "ample" have been updated and a test for the odreex::
ample::charseq<...> class template has been added as the "hello, world"
message in the form of a template "metastring".