Implemented map high order function for ppmpf sequences.
This is the version of map for ppmpf sequences. Both PPMPF_TUP_MAP and
PPMPF_SEQ_MAP result now in a finite series of preprocessor tokens
and each result from applying the function on either of the two ppmpf
collections is compliant to the following:

    PPMPF_TUP_MAP(f,(1,2,3))   -> ( (f(1)) , (f(2)) , (f(3)) )
    PPMPF_SEQ_MAP(f,(1)(2)(3)) -> (f(1)) (f(2)) (f(3))

The reason for having an additional enclosure on PPMPF_TUP_MAP has to do with
some ppmpf tuple operations being more vulnerable to special characters like
< > , [ ] { } etc. Sequences have no such problem. As a consequence, tuples
that are the result of PPMPF_TUP_MAP operations, have all their items
enclosed by default. Additional versions of map will be introduced to both
ppmpf collections to deal with this fact, as a prequel to "typed" versions of
this constructs coming about.