Added oprt metahandlers for ample type collections.
While the ppmpf component of odreex is about preprocessor metaprogramming, the
ample component (another metaprogramming library etude) is about C++ template
metaprogramming. Typelists are a fundamental C++ template metaprogramming
construct, with users of that style ending up in using or implementing various
kinds of type containers. Ample's type containers, to be referred to as
collections, are going to behave like the widely known C++ typelist constructs
with a few twists that are to reflect their differences with other solutions.

The <odreex/ample/oprt/storage.hh> header contains oprt metahandlers which are
the core templates ample algorithms are to deploy. A prior example of such
concept is found in the <odreex/ample/oprt/fundamentals.hh> header.

Given the load of different constructs both in both template and preprocessor
metaprogramming constructs, several branches in the odreex git repo are to
be introduced once ample's collection near completion.