Changing format flag in printf used in tests code.
In <odreex/ample/test.hh> test framework code, the printf function is used
with a "[%06zu]" format string containing the z flag which is an ISO C99
compliant option, but not an ISO C++11 one. Therefore and in order to continue
to use std::size_t portably with printf, switching to casting to unsigned long
long and deploying the [%06llu] format string is one of the options considered
for resolving this issue and continue using printf (for the time being...).

Thanks to Jose Diego Ferreira Martins <jose.diego@gmail.com> for reporting
this while compiling with mingw (4.7.2) on Microsoft Windows.