Using printf format specifiers in tests.
C++11 compliance includes C99 compliance, a fact that is at times left aside
when it comes to certain compilation environments and implementations of
standard libraries in Microsoft Windows. Since printf format specifiers are
used in the tests for the odreex library, the path of minimal change has been
selected as a way to portably "printf" integers used in tests code.

Tests compile regularly in *really* C++11 compliant environments with the
following options: -Wall -Wextra -pedantic. %lu compatible downcasting was
not necessary for this because C++11 and C99 have %llu. This change is
limited to the tests code and tests code alone. All other parts are
unaffected.