Expand __VA_ARGS__ properly in PPMPF_*F0 when extra arguments are used.
When certain amounts of arguments were used with placeholders in the folding
construct backend, due to PPMPF_*F0 not expanding properly, the placeholder
handler would not be passed along. This did not affect functions relying on
the original folding construct mechanism without the use of placeholders,
since expansion of the variadic arguments in PPMPF_*F0 was not necessary
for them. A really subtle error caught during review.