About autogen.sh.
The autogen.sh script is a GNU bash script that exists in order to assist with
autotools related code boilerplate, added some clarifications for end users
who may not be aware of what autogen.sh purpose is. It is unrelated to the C++
code in the odreex library and exists only for assisting autotools when
developers use fresh git clones of the repository.