Introduced ppmpf safe tuple constructor, resolved #8
This is the ppmpf safe tuple constructor, the following rules are valid:

    1) PPMPF_TUPLE()      -> ()
    2) PPMPF_TUPLE(1)     -> ((1))
    3) PPMPF_TUPLE(1,2)   -> ((1),(2))
    4) PPMPF_TUPLE(1,[],) -> ((1),([]))

Notice that if the last element used is a token with special characters, the
trailing comma is used as an escape character. Special characters are all
those that if concatenated with a random character, they result in at least
one invalid token. Tests have been added as well.