Added timers, make targets: ctim, jtim, tctime.
It is now possible to get the time required to compile a test with "make ctim"
or for all with "make tctime", while the timed version of "make just" is the
"make jtim" target. Errors are exported always to an errors.log file, as well
as times when asked for, in the form of CSV entries.