Several changes for more reusable ppmpf tests (#13).
Some of the purposes of issue #13 in the github issue tracker for odreex are
better dealt without increasing dependency on the autotools build system that
is currently deployed by the odreex library. This however involves additional
steps like providing an assistive build tool, coded within odreex itself.

The first step towards definitely resolving #13 is to refactor tests files
in such a way as to make them easily writable (and parsable) as well as
having better header coherence. In this commit:

    *) PPMPF_TST and PPMPF_TEST_MRUN are introduced, for the scope of
       deprecating other test macros.
    *) Gradual move towards more extensive PPMPF_TST use (see seq_foldr.cc)
    *) Some fixes and additions in <odreex/ample/charseq.hh>.
    *) run_all() introduced, check class template in <odreex/ample/test.hh>.
    *) A single header <odreex/test.hh> is to deal with testing dependencies.

The PPMPF_TST macro can eventually be rewired to control macro expansion while
compiling the tests, thus making use of the delayed expansion of the macro
function it is supposed to be creating the code for testing. All tests should
be converted to using it, new tests should deploy it.