Towards consolidating ppmpf tuple semantics (#18).
A series of changes that make the implementation of high order function macros
for both safe and unsafe ppmpf tuples cohere. Fixes have been applied to some
misbehaving PPMPF_UTUP_* members, and the *_1EMPTY and *_1GET functions got
properly renamed to the appropriate UTUP / TUP counterparts. Most importantly
PPMPF_UTUP_JOIN and PPMPF_UTUP_MAP have been properly adjusted and easily
participating from now one in compositions.

It is important to remember the two simple requirements for using unsafe ppmpf
tuples:

    1) If the final item in said unsafe tuples is a valid macro identifier or
       a non alphanumeric character sequence, a comma should immediately
       follow it during definition of the unsafe tuple.
    2) While combinations of the unsafe variants for ppmpf tuples is allowed
       by design, there should be a guarantee by the user that the first rule
       is not violated.

Safe variants do not need to comply to these two requirements. The reason for
maintaining safe and unsafe variants is that they are both doable with little
effort with the new recursive expansion constructs, while the unsafe variants
are of help in a series of algorithms that have particular needs when it comes
to performance at the preprocessor level.