A "type" system for the preprocessor using ppmpf typeclauses.
In relation to issue #6, ppmpf typeclauses refer to a 2-tuple of items
with the first being any kind of ppmpf collection, the second a decimal
number serving as a "type" identifier and referred to as typeid. This
concept is used in order to implement "overloadable" high order function
macros used in ppmpf internals, but it is more a convenient bypass than
anything else.