Prototypical macro function metahandlers for PPMPF_TFY in PPMPF_TMAP.
In order to pave the way for ppmpf expressions like "bind" to be used in
PPMPF_TFY controlled expressions, the function metahandlers to which
PPMPF_TFY delegates control of expansion should be the ones controlling
the expansion of a given function macro. As an example, PPMPF_TMAP_H1
and PPMPF_TMAP_H2 test handlers are rewired to detect a parenthesis in
place of a function as an "expression" and expanding to PPMPF_UNIT()
whenever that happens. In essence:

   1) Initially, the typeid is retrieved from the ppmpf typeclause
      containing the typeid and ppmpf collection data section through
      PPMPF_TFY.
   2) PPMPF_TFY dispatches to the right metahandler according to typeid
      retrieved from the ppmpf typeclause.
   3) Metahandlers (like PPMPF_TMAP_H0 and PPMPF_TMAP_H1) take over from
      PPMPF_TFY and dispatch to either the current implementations for
      their respective typeids or to PPMPF_UNIT in case an "expression"
      (right now just a fake one represented by the () token).

Practically, this means that the final effect of PPMPF_TFY is function
"transformation" and "type" manipulation fot whatever weight these
terms may carry at the preprocessor level. An interesting, functional
construct for essentially solving issues #6 and #20!