PPMPF_MAP is feature complete.
This commit deprecates all other map prototypes (though these may lie
around during refactoring period) and it bundles expressions, a "type"
system for the preprocessor based on ppmpf typeclauses in single macro
identifier. Both ppmpf sequences and tuples (safe and unsafe) now have
typeclauses that can be used to dispatch to the appropriate versions
of the macros involved in order to yield the effect of map as a high
order function. Closes #6, closes #18, closes #20. The same internals
are going to be used for all other high order functions already inside
ppmpf.