Deprecating <odreex/ppmpf/kernel/cpro/empty.hh>.
The reason behind this is that checking for the empty preprocessor token
is one of the most essential mechanisms for doing proper emulation of
recursive application of function macros over sequences of preprocessor
tokens. Thus, it has moved to <odreex/ppmpf/kernel/cpro/core.hh> and got
documented according to the internal documentation standard used in
odreex.