Several changes, project goes to MPL 2.0 licensing, named as 'clause'.
This commit resolves #19, which was about adding an exception to the GPLv3+
license used so far in this library. It was the explicit intention of the
author to begin with, that an exception should be added in the spirit of the
GCC linking exception to allow use by third party libraries that were not
under the GPL. The problem with this is that meta-programming techniques are
extremely hard to confine when it comes to licensing issues, because they
systematically produce code using various techniques and injecting it into
user codebases, therefore creating licensing issues depending on context.

Given that I am the sole author of this library and no releases have yet been
made, I decided to switch to Mozilla Public License version 2.0. The rationale
behind this is that it is an opensource license approved by both the OSI and
the FSF, it can be used with other licences whether 'closed' or 'open' but
most of all, there are no issues created when it comes to meta-programming.

Another important change is that this library is now called 'clause'. That
word seems to be more fitting for many reasons; to begin with, a clause is
the smallest grammatical unit that can express a complete proposition. That
is what 'clause' does for C++ meta-programs. It offers the tools for creating
complex metaprograms under of more 'basic' constructs.

Asides the project name and license switch to MPLv2, no other changes were
made.