Added ample::whenever for handling predicate bound cascades.
A logical disjunction over a sequence of predicate/action clause pairs.
Each such pair is evaluated and the first one to be true has its action
type parameter applied over the `oprt_apply metahandler embedded in the
metafunction itself to which a type parameter is passed.

The predicate/action clause pairs have to have ample::case_of compatible
semantics, while the resulting metafunction out of applying this to the
ample::case_of sequence is of course unary.