Adding a testcase for clause::ample::whenever.
The following example code, will instantiate to action9<temp__> because
the first case_of clause that evaluates to a true value is the one
where the action9 template type parameter is wrapper around an employ
template.

   apply<whenever< case_of<always_false, employ<action1> >
                 , case_of<always_false, employ<action2> >
                 , case_of<always_false, employ<action3> >
                 , case_of<always_false, employ<action4> >
                 , case_of<always_false, employ<action5> >
                 , case_of<always_false, employ<action6> >
                 , case_of<always_false, employ<action7> >
                 , case_of<always_false, employ<action8> >
                 , case_of<always_true , employ<action9> >, temp__ > >

In order to perform the test, ppmpf was used to create the boilerplate
for the various different cases to be checked in tests/ample/whenever.cc
while case_of and employ - being connected to whenever - are checked
as well.