Adding parsing mechanicms through TMP as clause::parse.
One of the known uses of template metaprogramming is for producing parsers
in a systematic manner. There are already several third party solutions to
this that deploy template metaprogramming techniques, this one is the
base to use within 'clause', it works for C++11 and up.

The user must be very familiar with parser theory, functional composition,
the work of B. Ford on PEGs (http://www.bford.info/pub/lang/peg.pdf, things
like recursive descent etc.

The clause::parse namespace is considered to be in constant flux for many
reasons, and it is to be used as the basis for solving issue #22, meaning
for generating the documentation of the library itself out of the comments,
provided I have some time for this. A simple example of use is found in the
'toys' subfolder of the clause project.