Introducing annotated template parameter packs in clause::ample.
Packs should be manipulated through the combination of integral constant
expressions in a series of primitives; this should be a language feature
in any self-respecting programming language dealing with the subject at
the compile-time level, especially when type-driven pattern matching is
widely deployed as in C++ template metaprogramming. Following primitives
have been implemented through the 'atpp' class:

 1) atpp<X...>::repeat<N>        // repeat X... N times
 2) atpp<X...>::pattern_of<N>    // X... is repetition of first N types
 3) atpp<X...>::expand<N,M>      // [N,M) expansion
 4) atpp<X...>::generate<N,M,K>  // [N,M) expansion * K times repetition
 5) atpp<X...>::pattern_is<Y...> // constituting pattern of X... is Y...
 6) atpp<X...>::atpos_is<N,T>    // at index N, we have type T
 7) atpp<X...>::atpos<N>         // get the type at index N
 8) atpp<X...>::insert<N,T>      // at index N insert type T
 9) atpp<X...>::insert<N,T...>   // at index N insert pack T...
10) atpp<X...>::replace<N,T...>  // T... replaces [N, N + sizeof...(T))

The whole idea can be used to base a series of compile-time type vectors
and other constructs and will be deployed in a certain fashion for such
purposes (asides the semantics study of ample::vector and the high order
functions deployed there) during development of 'clause'.