Getting sorted and unique type packs by alignment.
Forward declarations for several utility templates that are to be used
for emulating high order functions over a pack without using the rest
of the clause::ample library for two reasons: (1) alignof allows for
special optimizations when applied over a pack of types range for the
"sorting by alignof" step; (2) a single header solution was required
and depending on other parts would mean bring increasing compoments
of full-fledged metaprogramming library features. This header is to
provide utilities for aligned storage and it came up when a challenge
was thrown to me during a discussion with my fellow C++ programmers
Jonathan Müller and Manu Sánchez. Purpose of inclusion to `clause is
simple: it can be of use when analyzing boilerplate generation for
runtime containers and memory allocators by template metaprogramming.

tl;dr: fully standalone header for getting a duplicate-free, sorted by
alignment list of types unique by alignment.