Integral constant expressions in parameter pack manipulation.
Several changes involving to annotated template parameter packs as
described in issue #31. In essense this means that code like this:

template<typename...> struct test {};

// T... must be between 0 and 20 types in size while
// consisting of a repeating pattern of its first half;
// repeat it 8 times and instantiate a Super<...> class
// template with type indexes 1,3,5,0,0,7 then create
// a pack from of indices [3,5) of the resulting pack.
// Stranger things can happen.

template<typename... T>
auto example1(T...)
  -> templify((test) ( (T...){0,20}[2]^8)(1,3,5,0,0,7) >>= {3,5} );

is completely valid C++; there is a lot of documentation to be written
on how to actually make use of this but I need to get it running out
in the wild first so that it becomes easier to develop. It may be
that 'clause' itself may become entirely based on something like this.