Tests/examples where pattern factor operator is used for conjunction.
One of the consequences of the semantics of the pack operator is that
given that boolean expressions degenerate to 0/1 through implicit
conversions, we can exploit this fact to provide logical conjunction
in a series of cases, with multiple pattern factor operators expanded
upon different properties in the pack. SFINAE is triggered accordingly.

template<typename... T>
auto e12(T...)
  -> decltype( templify((std::tuple) (T...){}[true][sizeof...(T) > 1])()
             , a1() );